Split tests/test_adapaters_hermes_common.py tests into those for the common/hermes and common/utils tests
